+++ src/Symfony/Component/Validator/CHANGELOG.md
@@ -4,7 +4,7 @@ CHANGELOG
 8.2
 ---
 
- * Add the `PhoneNumber` constraint to validate phone numbers in the E.164 format, with an optional strict mode backed by the "giggsey/libphonenumber-for-php" library
+ * Add the `PhoneNumber` constraint to validate E.164-formatted phone numbers, with an optional strict mode backed by the "giggsey/libphonenumber-for-php" library
 
 8.1
 ---